arena.h: BREAKING CHANGE: ali functions take a reference to a list if they modify them.
Previously, if you had a list of type 'Foo **' (a list of pointers to Foo)
incorrect code like `ali_set_length(foo_list, 1);` where you forget the &
would compile and then fail at run time when the code checks the "magic"
code in the header.

The compiler could not tell that 'Foo **' is a list rather than a pointer
to a list at runtime.

Now you don't have to worry about forgetting the &: every macro
that uses a list takes the same thing as its first argument.